pyproject.toml
[project]
name = "sonarr-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastmcp>=3.1.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
